requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ftfy
regex
tqdm
Pillow>=8.0.0
numpy>=1.21.0
matplotlib>=3.3.0
git+https://github.com/openai/CLIP.git